# Conflicts:
#	projects/vampirism-api/src/main/java/de/teamlapen/vampirism/api/VEnums.java
#	projects/vampirism/src/main/resources/META-INF/enumextensions.json

+++ projects/vampirism/src/main/java/de/teamlapen/vampirism/common/world/blocks/WeaponTableBlock.java
@@ -160,7 +160,7 @@ public class WeaponTableBlock extends BaseHorizontalBlock {
 
             BlockState blockState = location.getBlockState();
 
-            return blockState.getValue(WeaponTableBlock.LAVA) * MB_PER_META;
+            return (long) blockState.getValue(WeaponTableBlock.LAVA) * MB_PER_META;
         }
 
         @Override
